pyproject.toml
[tool.poetry]
name = "crickformers"
version = "0.1.0"
description = "T20 ball-by-ball cricket data prediction"
authors = ["Shamus Rae <shamus@phi1618.io>"]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.0"
pytest = "^8.4.1"
numpy = "^2.0.1"
torch = "^2.5.1"
networkx = "^3.4.1"
torch-geometric = "^2.6.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 